+++ .github/workflows/ci.yml
@@ -1,0 +1,19 @@
+name: CI
+on: [push, pull_request]
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node: ['14', '12', '10']
+      name: Node ${{ matrix.node }}
+      steps:
+        - uses: actions/checkout@v1
+        - name: Setup node
+          uses: actions/setup-node@v1
+          with:
+            node-version: ${{ matrix.node }}
+        - run: npm install
+        - run: npm run standard
+        - run: npm run test
