--- test/types/tsconfig.json
@@ -1,23 +1,0 @@
-{
-  "files": ["../../fastify-auth.d.ts"],
-  "compilerOptions": {
-    "target": "es2017",
-    "lib": ["es2017"],
-    "skipLibCheck": true,
-    "baseUrl": "./",
-    "moduleResolution": "node",
-    "resolveJsonModule": true,
-    "checkJs": true,
-    "module": "commonjs",
-    "allowJs": true,
-    "noEmit": true,
-    "alwaysStrict": true,
-    "noImplicitAny": true,
-    "typeRoots": ["node_modules/@types"],
-    "allowSyntheticDefaultImports": true
-  },
-  "exclude": ["node_modules", "**/node_modules/*"],
-  "typeAcquisition": {
-    "enable": true
-  }
-}
